Moved SettingsObject to Settings Folder
