+++ src/Retail/Engine/EffectFormulas/Generic/TrinketDataTests.test.js
@@ -1,6 +1,6 @@
 
 import { getTrinketEffect, getProcessedValue } from "./TrinketEffectFormulas";
-import { userSettings } from "General/Modules/TopGear/SettingsObject";
+import { userSettings } from "General/Modules/Settings/SettingsObject";
 import Player from "General/Modules/Player/Player";
 import { trinket_data} from "./TrinketData";
 import each from "jest-each";
